Reorganize util code to work with jsdoc correctly
