+++ src/util/distanceGrid.js
@@ -50,150 +50,149 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 /**
- * @file
  * Code taken from https://github.com/Leaflet/Leaflet.markercluster
  * to support faster hierarchical clustering of features.
  * @copyright 2012, David Leaver
  */
 
-(function () {
-    "use strict";
+var $ = require('jquery');
+var L = {};
+L.Util = {
+    // return unique ID of an object
+    stamp: function (obj) {
+        obj._leaflet_id = obj._leaflet_id || ++L.Util.lastId;
+        return obj._leaflet_id;
+    },
+    lastId: 0
+};
 
-    var $ = require('jquery');
-    var L = {};
-    L.Util = {
-        // return unique ID of an object
-        stamp: function (obj) {
-            obj._leaflet_id = obj._leaflet_id || ++L.Util.lastId;
-            return obj._leaflet_id;
-        },
-        lastId: 0
-    };
+/**
+ * @class
+ * @alias geo.util.DistanceGrid
+ */
+var DistanceGrid = function (cellSize) {
+    this._cellSize = cellSize;
+    this._sqCellSize = cellSize * cellSize;
+    this._grid = {};
+    this._objectPoint = {};
+};
 
-    var DistanceGrid = function (cellSize) {
-        this._cellSize = cellSize;
-        this._sqCellSize = cellSize * cellSize;
-        this._grid = {};
-        this._objectPoint = {};
-    };
+DistanceGrid.prototype = {
 
-    DistanceGrid.prototype = {
+    addObject: function (obj, point) {
+        var x = this._getCoord(point.x),
+            y = this._getCoord(point.y),
+            grid = this._grid,
+            row = grid[y] = grid[y] || {},
+            cell = row[x] = row[x] || [],
+            stamp = L.Util.stamp(obj);
 
-        addObject: function (obj, point) {
-            var x = this._getCoord(point.x),
-                y = this._getCoord(point.y),
-                grid = this._grid,
-                row = grid[y] = grid[y] || {},
-                cell = row[x] = row[x] || [],
-                stamp = L.Util.stamp(obj);
+        point.obj = obj;
+        this._objectPoint[stamp] = point;
 
-            point.obj = obj;
-            this._objectPoint[stamp] = point;
+        cell.push(obj);
+    },
 
-            cell.push(obj);
-        },
+    updateObject: function (obj, point) {
+        this.removeObject(obj);
+        this.addObject(obj, point);
+    },
 
-        updateObject: function (obj, point) {
-            this.removeObject(obj);
-            this.addObject(obj, point);
-        },
+    //Returns true if the object was found
+    removeObject: function (obj, point) {
+        var x = this._getCoord(point.x),
+            y = this._getCoord(point.y),
+            grid = this._grid,
+            row = grid[y] = grid[y] || {},
+            cell = row[x] = row[x] || [],
+            i, len;
 
-        //Returns true if the object was found
-        removeObject: function (obj, point) {
-            var x = this._getCoord(point.x),
-                y = this._getCoord(point.y),
-                grid = this._grid,
-                row = grid[y] = grid[y] || {},
-                cell = row[x] = row[x] || [],
-                i, len;
+        delete this._objectPoint[L.Util.stamp(obj)];
 
-            delete this._objectPoint[L.Util.stamp(obj)];
+        for (i = 0, len = cell.length; i < len; i++) {
+            if (cell[i] === obj) {
 
-            for (i = 0, len = cell.length; i < len; i++) {
-                if (cell[i] === obj) {
+                cell.splice(i, 1);
 
-                    cell.splice(i, 1);
+                if (len === 1) {
+                    delete row[x];
+                }
 
-                    if (len === 1) {
-                        delete row[x];
+                return true;
+            }
+        }
+
+    },
+
+    eachObject: function (fn, context) {
+        var i, j, k, len, row, cell, removed,
+            grid = this._grid;
+
+        for (i in grid) {
+            row = grid[i];
+
+            for (j in row) {
+                cell = row[j];
+
+                for (k = 0, len = cell.length; k < len; k++) {
+                    removed = fn.call(context, cell[k]);
+                    if (removed) {
+                        k--;
+                        len--;
                     }
-
-                    return true;
                 }
             }
+        }
+    },
 
-        },
+    getNearObject: function (point) {
+        var x = this._getCoord(point.x),
+            y = this._getCoord(point.y),
+            i, j, k, row, cell, len, obj, dist,
+            objectPoint = this._objectPoint,
+            closestDistSq = this._sqCellSize,
+            closest = null;
 
-        eachObject: function (fn, context) {
-            var i, j, k, len, row, cell, removed,
-                grid = this._grid;
+        for (i = y - 1; i <= y + 1; i++) {
+            row = this._grid[i];
+            if (row) {
 
-            for (i in grid) {
-                row = grid[i];
-
-                for (j in row) {
+                for (j = x - 1; j <= x + 1; j++) {
                     cell = row[j];
+                    if (cell) {
 
-                    for (k = 0, len = cell.length; k < len; k++) {
-                        removed = fn.call(context, cell[k]);
-                        if (removed) {
-                            k--;
-                            len--;
-                        }
-                    }
-                }
-            }
-        },
-
-        getNearObject: function (point) {
-            var x = this._getCoord(point.x),
-                y = this._getCoord(point.y),
-                i, j, k, row, cell, len, obj, dist,
-                objectPoint = this._objectPoint,
-                closestDistSq = this._sqCellSize,
-                closest = null;
-
-            for (i = y - 1; i <= y + 1; i++) {
-                row = this._grid[i];
-                if (row) {
-
-                    for (j = x - 1; j <= x + 1; j++) {
-                        cell = row[j];
-                        if (cell) {
-
-                            for (k = 0, len = cell.length; k < len; k++) {
-                                obj = cell[k];
-                                dist = this._sqDist(
-                                    objectPoint[L.Util.stamp(obj)],
-                                    point
-                                );
-                                if (dist < closestDistSq) {
-                                    closestDistSq = dist;
-                                    closest = obj;
-                                }
+                        for (k = 0, len = cell.length; k < len; k++) {
+                            obj = cell[k];
+                            dist = this._sqDist(
+                                objectPoint[L.Util.stamp(obj)],
+                                point
+                            );
+                            if (dist < closestDistSq) {
+                                closestDistSq = dist;
+                                closest = obj;
                             }
                         }
                     }
                 }
             }
-            return closest;
-        },
-
-        /* return the point coordinates contained in the structure */
-        contents: function () {
-            return $.map(this._objectPoint, function (val) { return val; });
-        },
-
-        _getCoord: function (x) {
-            return Math.floor(x / this._cellSize);
-        },
-
-        _sqDist: function (p, p2) {
-            var dx = p2.x - p.x,
-                dy = p2.y - p.y;
-            return dx * dx + dy * dy;
         }
-    };
+        return closest;
+    },
 
-    module.exports = DistanceGrid;
-})();
+    /* return the point coordinates contained in the structure */
+    contents: function () {
+        return $.map(this._objectPoint, function (val) { return val; });
+    },
+
+    _getCoord: function (x) {
+        return Math.floor(x / this._cellSize);
+    },
+
+    _sqDist: function (p, p2) {
+        var dx = p2.x - p.x,
+            dy = p2.y - p.y;
+        return dx * dx + dy * dy;
+    }
+};
+
+module.exports = DistanceGrid;
